+++ packages/cloud_functions/cloud_functions/lib/src/cloud_functions.dart
@@ -17,7 +17,7 @@ class CloudFunctionsException implements Exception {
 /// You can get an instance by calling [CloudFunctions.instance].
 class CloudFunctions {
   CloudFunctions({FirebaseApp app, String region})
-      : _app = app ?? FirebaseApp.instance,
+      : _app = app ?? Firebase.app(),
         _region = region;
 
   static CloudFunctions _instance = CloudFunctions();