+++ packages/firebase_admob/example/android/app/src/main/java/io/flutter/plugins/firebaseadmobexample/MainActivity.java
@@ -7,6 +7,7 @@ package io.flutter.plugins.firebaseadmobexample;
 import dev.flutter.plugins.e2e.E2EPlugin;
 import io.flutter.embedding.android.FlutterActivity;
 import io.flutter.embedding.engine.FlutterEngine;
+import io.flutter.plugins.firebase.core.FlutterFirebaseCorePlugin;
 import io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin;
 import io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin.NativeAdFactory;
 
@@ -16,6 +17,7 @@ public class MainActivity extends FlutterActivity {
     // TODO(bparrishMines): Remove this once v2 of GeneratedPluginRegistrant rolls to stable. https://github.com/flutter/flutter/issues/42694
     flutterEngine.getPlugins().add(new E2EPlugin());
     flutterEngine.getPlugins().add(new FirebaseAdMobPlugin());
+    flutterEngine.getPlugins().add(new FlutterFirebaseCorePlugin());
 
     final NativeAdFactory factory = new NativeAdFactoryExample(getLayoutInflater());
     FirebaseAdMobPlugin.registerNativeAdFactory(flutterEngine, "adFactoryExample", factory);